+++ src/Common/utils.ts
@@ -5,7 +5,7 @@ import * as addressCodec from 'ripple-address-codec'
 /**
  * A prefix applied when hashing a signed transaction blob.
  *
- * @see https://xrpl.org/basic-data-types.html#hashes
+ * @see {@link https://xrpl.org/basic-data-types.html#hashes|XRPL Data Types - Hashes}.
  */
 const signedTransactionPrefixHex = '54584E00'
 
@@ -27,11 +27,12 @@ export interface ClassicAddress {
 const utils = {
   /**
    * Validate that the given string is a valid address for the XRP Ledger.
-   *
    * This function returns true for both X-addresses and classic addresses.
-   * @see https://xrpaddress.info/
+   *
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
    *
    * @param address - An address to check.
+   *
    * @returns True if the address is valid, otherwise false.
    */
   isValidAddress(address: string): boolean {
@@ -44,7 +45,7 @@ const utils = {
   /**
    * Validate whether the given string is a valid X-address for the XRP Ledger.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
    *
    * @param address - An address to check.
    * @returns True if the address is a valid X-address, otherwise false.
@@ -56,7 +57,7 @@ const utils = {
   /**
    * Validate whether the given string is a valid classic address for the XRP Ledger.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
    *
    * @param address - An address to check.
    * @returns True if the address is a valid classic address, otherwise false.
@@ -68,7 +69,7 @@ const utils = {
   /**
    * Encode the given classic address and tag into an x-address.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
    *
    * @param classicAddress - A classic address to encode.
    * @param tag - An optional tag to encode.
@@ -96,10 +97,10 @@ const utils = {
   /**
    * Decode a `ClassicAddress` from a given x-address.
    *
-   * @see https://xrpaddress.info/
+   * @see {@link https://xrpaddress.info/|XRPL X-Address Format}.
    *
    * @param xAddress - The xAddress to decode.
-   * @returns A `ClassicAddress`
+   * @returns A `ClassicAddress`.
    */
   decodeXAddress(xAddress: string): ClassicAddress | undefined {
     if (!addressCodec.isValidXAddress(xAddress)) {
@@ -118,7 +119,7 @@ const utils = {
   /**
    * Convert the given byte array to a hexadecimal string.
    *
-   * @param bytes - An array of bytes
+   * @param bytes - An array of bytes.
    * @returns An encoded hexadecimal string.
    */
   toHex(bytes: Uint8Array): string {
@@ -159,7 +160,7 @@ const utils = {
    * Check if the given string is valid hex.
    *
    * @param input - The input to check.
-   * @returns true if the input is valid hex, otherwise false.
+   * @returns True if the input is valid hex, otherwise false.
    */
   isHex(input: string): boolean {
     const hexRegEx = /(?:[0-9]|[a-f])/gimu
@@ -169,7 +170,7 @@ const utils = {
   /**
    * Compute the SHA512 half hash of the given bytes.
    *
-   * @param input - The input to hash.
+   * @param bytes - The array of bytes to hash.
    * @returns The hash of the input.
    */
   sha512Half(bytes: Uint8Array): Uint8Array {